deploy/main_deploy/requirements.txt
beartype==0.22.9
certifi==2026.2.25
charset-normalizer==3.4.5
Faker==40.1.0
idna==3.11
pyaml==26.2.1
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
tzdata==2025.3
urllib3==2.6.3
prefect>=3.0.0
pydantic>=2.0.0
pyarrow>=15.0.0
pandas>=2.2.0
loguru>=0.7.0
faker>=25.0.0
pyspark==3.5.3
delta-spark==3.2.1